# api/openapi.yaml
openapi: 3.0.3
info:
  title: memoryd API
  version: 0.1.0
  description: |
    HTTP contract for the local Memory Vault. A Memory is one user-visible
    thing preserved by the Vault and is identified by the content of its Blob.
servers:
  - url: /api/v0
    description: Local memoryd server
...
paths:
...
  /memories:
    get:
      operationId: browseMemories
      tags: [memories]
      summary: Browse Memories
      description: Returns whole Memories in reverse original-time order.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of Memories.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryPage"
        "400":
          $ref: "#/components/responses/BadRequest"
  /memories/{memoryId}:
    parameters:
      - $ref: "#/components/parameters/MemoryId"
    get:
      operationId: getMemory
      tags: [memories]
      summary: Get Memory details
      responses:
        "200":
          description: The Memory, its Facts, Derived Content, and active Run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /memories/{memoryId}/content:
    parameters:
      - $ref: "#/components/parameters/MemoryId"
    get:
      operationId: getMemoryContent
      tags: [memories]
      summary: Download a Memory's immutable Blob
      responses:
        "200":
          description: The original imported bytes.
          headers:
            Content-Disposition:
              schema:
                type: string
              description: Suggested filename for download.
            ETag:
              schema:
                type: string
              description: The content hash, quoted as an HTTP entity tag.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
  /memories/import:
    post:
      operationId: importMemory
      tags: [memories]
      summary: Import one Blob and understand it
      description: |
        Upload one supported Blob. The request remains open until the
        understanding attempt reaches a terminal result. The response is an
        SSE stream; the client owns directory enumeration and uploads Blobs
        sequentially. Once the server has committed the Blob and InProgress
        Memory, disconnecting does not cancel server-side understanding.
        Blobs are limited to 100 MiB. Duplicate content is rejected with HTTP
        409 before an SSE response begins.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImportMultipart"
            encoding:
              file:
                contentType: application/pdf, image/jpeg, image/png
      responses:
        "200":
          description: Progress events followed by exactly one terminal event.
          content:
            text/event-stream:
              schema:
                type: string
                description: UTF-8 Server-Sent Events. Each data field is JSON.
              examples:
                progress:
                  summary: Understanding progress
                  value: |-
                    event: understanding_progress
                    data: {"memory_id":"2d6f4d1a-4d62-4ef3-9b2c-6aa7f1f0d6c2","phase":"general_extraction","message":"Extracting text"}

                completed:
                  summary: Successful terminal event
                  value: |-
                    event: import_completed
                    data: {"memory":{"id":"2d6f4d1a-4d62-4ef3-9b2c-6aa7f1f0d6c2","blob_hash":"0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef","media_type":"application/pdf","byte_size":1234,"original_filename":"bill.pdf","understanding_state":"Done","imported_at":"2026-01-01T00:00:00Z"}}
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: The candidate is a duplicate or an import cannot start.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: The request exceeds the server's upload limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: The Blob format is not supported by the MVP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /memories/{memoryId}/runs:
    parameters:
      - $ref: "#/components/parameters/MemoryId"
    get:
      operationId: listUnderstandingRuns
      tags: [understanding]
      summary: List successful Understanding Runs
      responses:
        "200":
          description: Runs for the Memory, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunPage"
        "404":
          $ref: "#/components/responses/NotFound"
  /memories/{memoryId}/runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/MemoryId"
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: getUnderstandingRun
      tags: [understanding]
      summary: Get one Understanding Run
      responses:
        "200":
          description: An immutable successful interpretation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnderstandingRun"
        "404":
          $ref: "#/components/responses/NotFound"
  /memories/{memoryId}/logs:
    parameters:
      - $ref: "#/components/parameters/MemoryId"
    get:
      operationId: listProcessingLogs
      tags: [understanding]
      summary: List processing logs for a Memory
      description: Includes logs for failed attempts that did not create a Run.
      parameters:
        - name: attempt_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 1000
            default: 200
      responses:
        "200":
          description: Processing log entries in lifecycle order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingLogPage"
        "404":
          $ref: "#/components/responses/NotFound"
  /memories/{memoryId}/runs/{runId}/logs:
    parameters:
      - $ref: "#/components/parameters/MemoryId"
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: listRunLogs
      tags: [understanding]
      summary: List logs associated with an Understanding Run
      responses:
        "200":
          description: Processing logs associated with the Run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingLogPage"
        "404":
          $ref: "#/components/responses/NotFound"
  /memories/{memoryId}/rebuild:
    parameters:
      - $ref: "#/components/parameters/MemoryId"
    post:
      operationId: rebuildMemory
      tags: [understanding]
      summary: Rebuild a Memory's understanding
      description: Stages a complete new Run and activates it atomically on success.
      responses:
        "200":
          $ref: "#/components/responses/UnderstandingEventStream"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A rebuild is already in progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /memories/{memoryId}/codex-enhancement:
    parameters:
      - $ref: "#/components/parameters/MemoryId"
    post:
      operationId: enhanceMemoryWithCodex
      tags: [understanding]
      summary: Explicitly request Codex-assisted enhancement
      description: |
        Creates a complete new Run containing the regular pipeline result plus
        Codex enrichment. The client must confirm before repeating an equivalent
        prior Codex enhancement.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodexEnhancementRequest"
      responses:
        "200":
          $ref: "#/components/responses/UnderstandingEventStream"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Confirmation is required before repeating an equivalent enhancement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: The Blob cannot be prepared within the cloud-bound input limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /search:
    post:
      operationId: searchMemories
      tags: [search]
      summary: Search Memories with a natural-language query
      description: |
        Interprets the query into a visible Query Plan and immediately executes
        exact Fact filters plus mandatory residual full-text terms.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Query Plan and whole-Memory search results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: The query could not be interpreted into a valid Query Plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    MemoryId:
      name: memoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested Memory or Run does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnderstandingEventStream:
      description: Progress events followed by one terminal rebuild or enhancement event.
      content:
        text/event-stream:
          schema:
            type: string
            description: UTF-8 Server-Sent Events. Each data field is JSON.
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: memory_not_found
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        existing_memory:
          allOf:
            - $ref: "#/components/schemas/DuplicateMemory"
          nullable: true
          description: Present on a duplicate import conflict so clients can recover.
    ImportMultipart:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          maxLength: 104857600
          description: One PDF, JPEG, or PNG Blob.
        relative_path:
          type: string
          description: Browser-provided path relative to the selected directory.
        full_path:
          type: string
          description: Original full path when the importing client can provide it.
        media_type:
          type: string
          description: Client-observed media type; the server validates the content.
        filesystem_created_at:
          type: string
          format: date-time
        filesystem_modified_at:
          type: string
          format: date-time
...
    MemorySummary:
      type: object
      required:
        - id
        - blob_hash
        - media_type
        - byte_size
        - original_filename
        - understanding_state
        - imported_at
      properties:
        id:
          type: string
          format: uuid
        blob_hash:
          type: string
          pattern: "^[a-fA-F0-9]{64}$"
          description: SHA-256 content hash of the immutable Blob.
        media_type:
          type: string
        byte_size:
          type: integer
          format: int64
          minimum: 0
        original_filename:
          type: string
        understanding_state:
          $ref: "#/components/schemas/UnderstandingState"
        active_run_id:
          type: string
          format: uuid
          nullable: true
        imported_at:
          type: string
          format: date-time
        original_created_at:
          type: string
          format: date-time
          nullable: true
        original_modified_at:
          type: string
          format: date-time
          nullable: true
    UnderstandingState:
      type: string
      enum: [InProgress, Done, Failed]
    MemoryPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MemorySummary"
        next_cursor:
          type: string
          nullable: true
    MemoryDetail:
      type: object
      required: [memory, import_context, facts, derived_content]
      properties:
        memory:
          $ref: "#/components/schemas/MemorySummary"
        content_url:
          type: string
          description: Relative URL of the content download endpoint.
        import_context:
          $ref: "#/components/schemas/ImportContext"
        facts:
          type: array
          items:
            $ref: "#/components/schemas/Fact"
        derived_content:
          type: array
          items:
            $ref: "#/components/schemas/DerivedContent"
        active_run:
          allOf:
            - $ref: "#/components/schemas/UnderstandingRun"
          nullable: true
    ImportContext:
      type: object
      required: [original_filename]
      properties:
        original_filename:
          type: string
        relative_path:
          type: string
        full_path:
          type: string
        media_type:
          type: string
        byte_size:
          type: integer
          format: int64
          minimum: 0
        content_hash:
          type: string
          pattern: "^[a-fA-F0-9]{64}$"
        filesystem_created_at:
          type: string
          format: date-time
          nullable: true
        filesystem_modified_at:
          type: string
          format: date-time
          nullable: true
...
    DerivedContent:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          example: extracted_text
        text:
          type: string
        metadata:
          type: object
          additionalProperties: true
    UnderstandingRun:
      type: object
      required: [id, memory_id, pipeline, created_at, completed_at, active]
      properties:
        id:
          type: string
          format: uuid
        memory_id:
          type: string
          format: uuid
        pipeline:
          type: string
          enum: [regular, codex_enhancement]
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        active:
          type: boolean
        extractor_versions:
          type: object
          additionalProperties:
            type: string
        plugin_versions:
          type: object
          additionalProperties:
            type: string
        warnings:
          type: array
          items:
            type: string
    RunPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UnderstandingRun"
        next_cursor:
          type: string
          nullable: true
...
    ProcessingLogPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProcessingLog"
        next_cursor:
          type: string
          nullable: true
    CodexEnhancementRequest:
      type: object
      properties:
        instructions:
          type: string
          description: Optional user direction for the explicit Codex task.
        confirm_repeat:
          type: boolean
          default: false
          description: Confirm repeating an equivalent prior Codex enhancement.
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        limit:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          default: 50
        cursor:
          type: string
    SearchResponse:
      type: object
      required: [query, plan, items]
      properties:
        query:
          type: string
        plan:
          $ref: "#/components/schemas/QueryPlan"
        items:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        next_cursor:
          type: string
          nullable: true
...